rename bindings -> fluent
