+++ fluent/Entry.ts
