+++ fluent/Signed.ts
