+++ mod.ts
@@ -1,5 +1,5 @@
 export * as $ from "./_deps/scale.ts";
-export * from "./bindings/mod.ts";
 export * from "./Config.ts";
+export * from "./fluent/mod.ts";
 export * as test from "./test-util/mod.ts";
 export { iter } from "./util/mod.ts";
