+++ runtime/read.ts
@@ -1,4 +1,4 @@
-import * as core from "../bindings/mod.ts";
+import * as core from "../fluent/mod.ts";
 import * as M from "../frame_metadata/mod.ts";
 import { Hashers } from "../hashers/mod.ts";
 import * as U from "../util/mod.ts";
